Implement Feature #78
Implement settings logic for feature in #78 including saving the new setting to the Settings.INI file. This will now make it so the user has to manually enable the automatic SMAPI check. It will be disabled by default as it can be a resource intensive operation and can lag slower hardware.

Co-Authored-By: Ryan Walpole <69621127+RyanWalpole@users.noreply.github.com>